instancetype: Extract builder out of tests
