+++ pkg/instancetype/builder/v1alpha1/BUILD.bazel
@@ -3,12 +3,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["builder.go"],
-    importpath = "kubevirt.io/kubevirt/tests/libinstancetype/builder",
+    importpath = "kubevirt.io/kubevirt/pkg/instancetype/builder/v1alpha1",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/pointer:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
-        "//staging/src/kubevirt.io/api/instancetype/v1beta1:go_default_library",
+        "//staging/src/kubevirt.io/api/instancetype/v1alpha1:go_default_library",
         "//tests/framework/cleanup:go_default_library",
         "//tests/testsuite:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
