+++ tests/instancetype/BUILD.bazel
@@ -13,6 +13,9 @@ go_library(
         "//pkg/apimachinery/patch:go_default_library",
         "//pkg/controller:go_default_library",
         "//pkg/instancetype:go_default_library",
+        "//pkg/instancetype/builder:go_default_library",
+        "//pkg/instancetype/builder/v1alpha1:go_default_library",
+        "//pkg/instancetype/builder/v1alpha2:go_default_library",
         "//pkg/instancetype/conflict:go_default_library",
         "//pkg/instancetype/revision:go_default_library",
         "//pkg/libdv:go_default_library",
@@ -30,9 +33,6 @@ go_library(
         "//tests/decorators:go_default_library",
         "//tests/framework/kubevirt:go_default_library",
         "//tests/framework/matcher:go_default_library",
-        "//tests/libinstancetype/builder:go_default_library",
-        "//tests/libinstancetype/builder/v1alpha1:go_default_library",
-        "//tests/libinstancetype/builder/v1alpha2:go_default_library",
         "//tests/libkubevirt:go_default_library",
         "//tests/libkubevirt/config:go_default_library",
         "//tests/libstorage:go_default_library",
